+++ bin/RUNA3_only_Z.csh
@@ -1,6 +1,6 @@
 #!/bin/csh -f
 #
-# W phase package - RUNA3 for Z components
+# W phase package - RUNA3 for Z, N, E components
 #
 # Zacharie Duputel, Luis Rivera and Hiroo Kanamori
 #
@@ -8,54 +8,67 @@
 source $WPHASE_HOME/bin/WP_HEADER.CSH
 ##################################
 
+if ( ! -e i_master ) then
+   ${ECHO} "Error: file  i_master not available"
+   exit 1
+endif
+
+set my_argv = ($ARGV)
 set BIN     = $WPHASE_HOME/bin
-set EXTRACT = ${BIN}/extract_only_Z.csh
-set CALC    = ${BIN}/calc_fast_synths_only_Z.csh
+set EXTRACT = ${BIN}/extract.csh
 set PREPARE = ${BIN}/prepare_wp_ZNE.csh
 set WPINVER = ${BIN}/wpinversion_ZNE
 
+set median  = "-med"
+set p2p_scr = `$GREP ^P2P_SCREENING i_master| $CUT -d':' -f2`
+if ( $#p2p_scr != 0 && $p2p_scr != "YES" ) set median = ""
+set ths     = "5.0 3.0 0.9"
+set rms_scr = `$GREP ^RMS_SCREENING i_master| $CUT -d':' -f2`
+if ( $#rms_scr != 0 ) set ths = "$rms_scr"
+ 
 
-${RM} -rf SYNTH
-${MKDIR} SYNTH
-
+if ( -e SYNTH ) ${RM} -rf SYNTH 
+${MKDIR} SYNTH; 
 $EXTRACT
-$CALC
-if $status exit(1)
-$PREPARE
+if $status exit 1
+$PREPARE Z
+if $status exit 1
 
-if (-e i_master) then
-        ${GREP} -v "^#" i_master >! i_tmp
-else
-        ${ECHO} "Error: file  i_master not available"
-        exit
-endif
-set gf_dir   = "./GF"
-set tmp      = `${GREP} GFDIR   i_tmp`
-if ! $status then
-        set gf_dir   = `$ECHO $tmp | ${HEAD} -1 | ${CUT} -d: -f2`
-endif
-${RM} -f i_tmp 
+set version = "Version: "
+if ( -e $WPHASE_HOME/.svn/entries ) set version = "${version}`$HEAD -4 $WPHASE_HOME/.svn/entries | $TAIL -1`"
+set screening = "Screening: $median"
+set comgfdir  = "GF_PATH: $GF_PATH"
 
-$WPINVER -log LOG/wpinversion.noth.log -osyndir SYNTH -gfdir ${gf_dir} \
-  	 -pdata fort.15.noth 
+$ECHO "COMMAND LINE: $WPINVER -log LOG/wpinversion.noth.log -osyndir SYNTH -pdata fort.15.noth $median \
+	${my_argv} -comment '$version' -comment '$screening' -comment '$comgfdir'"
+$WPINVER -log LOG/wpinversion.noth.log -osyndir SYNTH -pdata fort.15.noth $median \
+	${my_argv} -comment "$version" -comment "$screening" -comment "$comgfdir"
 
-${CP} p_wpinversion p_wpinversion.noth
+${CP} p_wpinversion.ps p_wpinversion.noth.ps
 ${CP} o_wpinversion o_wpinversion.noth
 ${CP} WCMTSOLUTION WCMTSOLUTION.noth
 
 ${CP} o_wpinversion o_wpinv
-set ths =  "5.0 3.0 0.9"
+
 
 foreach th ($ths)
-
+	set screening = "$screening  $th"
+    $ECHO -e "\nCOMMAND LINE: $WPINVER -th ${th} -ifil o_wpinversion -ofil o_wpinv.th_${th} \
+	-log LOG/wpinversion.th_${th}.log -ps p_wpinversion.th_${th} \
+	-osyndir SYNTH -ocmtf  WCMTSOLUTION.th_${th} -old ${my_argv} \
+	-comment '$version' -comment '$screening' -comment '$comgfdir'" 
     $WPINVER -th ${th} -ifil o_wpinversion -ofil o_wpinv.th_${th} \
-    -log LOG/wpinversion.th_${th}.log -ps p_wpinversion.th_${th} \
-    -osyndir SYNTH -ocmtf  WCMTSOLUTION.th_${th}\
-    -gfdir ${gf_dir} -old
-    if ($status == 1) exit(1)
-    ${CP} -f o_wpinv.th_$th o_wpinversion        
+    -log LOG/wpinversion.th_${th}.log -ps p_wpinversion.th_${th}.ps \
+    -osyndir SYNTH -ocmtf  WCMTSOLUTION.th_${th} -old ${my_argv} \
+	-comment "$version" -comment "$screening" -comment "$comgfdir" 
+    if $status exit 1
+    ${CP} -f o_wpinv.th_$th o_wpinversion
 end
 
 ${CP} WCMTSOLUTION.th_${th} WCMTSOLUTION
-${CP} p_wpinversion.th_${th} p_wpinversion
+${CP} p_wpinversion.th_${th}.ps p_wpinversion.ps
 ${CP} LOG/wpinversion.th_${th}.log LOG/wpinversion.log
+#${RM} -rf GF
+
+${ECHO} -e "\nOutput files: o_wpinversion WCMTSOLUTION p_wpinversion.ps"
+${ECHO} "              fort.15 fort.15_LHZ fort.15_LHN fort.15_LHE"
